requirements.txt
absl-py==1.4.0
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
clldutils==3.19.0
cmake==3.27.4.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.7.0
contourpy==1.1.0
csvw==3.1.3
cycler==0.11.0
Cython==3.0.2
dataclasses-json==0.5.14
decorator==5.1.1
dlinfo==1.2.1
exceptiongroup==1.1.3
executing==1.2.0
fastapi==0.103.1
ffmpeg-python==0.2.0
ffmpy==0.3.1
filelock==3.12.3
flatbuffers==23.5.26
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.0
future==0.18.3
google-auth==2.23.0
google-auth-oauthlib==1.0.0
gradio==3.43.2
gradio_client==0.5.0
greenlet==2.0.2
grpcio==1.58.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.1
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
ipython==8.15.0
isodate==0.6.1
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.286
langsmith==0.0.35
language-tags==1.2.0
lazy_loader==0.3
librosa==0.10.1
lit==15.0.7
llvmlite==0.40.1
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
oauthlib==3.2.2
onnx==1.14.1
onnxruntime==1.15.1
openai==0.28.0
openai-whisper==20230314
orjson==3.9.7
packaging==23.1
pandas==2.1.0
parso==0.8.3
pexpect==4.8.0
phonemizer==3.2.1
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
Pygments==2.16.1
pylatexenc==2.10
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rdflib==7.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rpds-py==0.10.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.2
segments==2.2.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.6
SQLAlchemy==2.0.20
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tiktoken==0.3.1
toolz==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
tqdm==4.66.1
traitlets==5.9.0
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.7
yarl==1.9.2
